requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==4.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
pillow==11.1.0
psycopg==3.2.6
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.1
